Third Commit

+++ src/main/kotlin/com/example/rest/repository/TodoRepo.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface TodoRepo: JpaRepository<TodoItem, Integer> {
+interface TodoRepo: JpaRepository<TodoItem, Long> {
 
 }
